do not post local-only :options values to amara.org endpoints
these are already stripped out of the top-level of `params`, but they will
instead exist in params[:data] during `Amara::API#create` and
`Amara::API#update` calls.

before this change

```
  api = Amara::API.new
  api.create(a:1, b:2)
```

would result in this post body:

```
{
  "a":1,
  "b":2,
  "options":{
    "raise_errors":true
  }
}
```

the 'options' object is ignored by amara.org, but this caused errors for me
when attempting to use VCR/webmock to match test requests based on the POST
contents being sent to the server.